refactor(test): move reader_constructor fixture to conftest
- Added new conftest file to store common reader_constructor fixture
- Removed reader_constructor fixtures from 5 files

Note: test_distributed_training.py and test_multiprocess_dataloading also could use READER_CONSTRUCTORS
variable in the future, but might require converting tst folder to module. Added TODO for now.
