+++ s3torchconnector/tst/unit/dcp/test_s3_file_system.py
@@ -26,20 +26,6 @@ TEST_DATA = b"test data\n"
 TEST_PATH = f"s3://{TEST_BUCKET}/{TEST_KEY}"
 
 
-@pytest.fixture(
-    params=[
-        S3ReaderConstructor.sequential(),  # Sequential Reader
-        S3ReaderConstructor.range_based(),  # range-based reader w/ buffer
-        S3ReaderConstructor.range_based(buffer_size=0),  # range-based reader, no buffer
-    ],
-    ids=["sequential", "range_based_buffered", "range_based_unbuffered"],
-    scope="module",
-)
-def reader_constructor(request) -> S3ReaderConstructorProtocol:
-    """Provide reader constructor (partial(S3Reader)) instances for all supported reader types."""
-    return request.param
-
-
 def test_create_stream():
     mock_client = MockS3Client(TEST_REGION, TEST_BUCKET)
     s3fs = S3FileSystem(TEST_REGION, mock_client)
